Update package.json
Some starter code was given by the tutor to help rework homework functionality. Code that was given apparently from another repo then was added to by Ben (tutor) and Josh (TA) when helped with the functionality of the assignment. Updated .json file to match the current repo after speaking with Lisa (SSM).